requirements.txt
attrs==18.2.0
matplotlib==3.0.3
numpy==1.16.0